Use frontend for uBPF
