+++ backends/ubpf/ubpfModel.h
@@ -99,15 +99,6 @@ namespace UBPF {
         int numberOfControlBlockArguments() { return version >= 20200515 ? 3 : 2; }
     };
 
-    class getUBPFModelVersion : public Inspector {
-        bool preorder(const IR::Declaration_Constant *dc) override {
-            if (dc->name == "__ubpf_model_version") {
-                auto val = dc->initializer->to<IR::Constant>();
-                UBPFModel::instance.version = static_cast<unsigned>(val->value); }
-            return false; }
-        bool preorder(const IR::Declaration *) override { return false; }
-    };
-
 }
 
 #endif
